requirements.txt
Django==1.10.2
django-filter==0.15.2
djangorestframework==3.4.7
djoser==0.5.1
gunicorn==19.6.0
Markdown==2.6.7
psycopg2==2.6.2
six==1.10.0
